feat(frontend): Batch 3 - extract features/referenceData/
Create features/referenceData/ and consolidate all reference-data
management files from features/classes/management/ and features/settings/:

From features/settings/:
  - ManageTopicsModal, ReferenceDataSettingsPanel

From features/classes/management/:
  - ManageCohortsModal, ManageYearGroupsModal
  - ReferenceDataManagementModalScaffold
  - manageReferenceDataDialogs, manageReferenceDataHelpers
  - All associated spec files

Update imports in:
  - ClassesManagementPanel, useReferenceDataManagement
  - AssignmentDefinitionWizardModal
  - SettingsPage and related test helpers

Remove now-empty features/classes/management/ directory.

Lint: clean. Tests: 91/91 pass, 894 tests.
