+++ src/frontend/src/features/assignmentWizard/AssignmentDefinitionWizardModal.tsx
@@ -2,8 +2,8 @@ import { Modal, Space, Button, Typography } from 'antd';
 import { useCallback, useState, type JSX } from 'react';
 import { useAssignmentDefinitionWizard } from './useAssignmentDefinitionWizard';
 import { AssignmentDefinitionWizardModalShell } from './AssignmentDefinitionWizardModalShell';
-import { ManageTopicsModal } from '../../features/settings/ManageTopicsModal';
-import { ManageYearGroupsModal } from '../../features/classes/management/ManageYearGroupsModal';
+import { ManageTopicsModal } from '../referenceData/ManageTopicsModal';
+import { ManageYearGroupsModal } from '../referenceData/ManageYearGroupsModal';
 import { type AssignmentDefinitionWizardModalProperties } from './useAssignmentDefinitionWizard';
 
 const { Text } = Typography;
