+++ src/frontend/src/features/classes/ClassesManagementPanel.tsx
@@ -13,8 +13,8 @@ import { BulkCreateModal } from './bulk/BulkCreateModal';
 import { BulkDeleteModal } from './bulk/BulkDeleteModal';
 import { BulkSetCourseLengthModal } from './bulk/BulkSetCourseLengthModal';
 import { BulkSetSelectModal } from './bulk/BulkSetSelectModal';
-import { ManageCohortsModal } from './management/ManageCohortsModal';
-import { ManageYearGroupsModal } from './management/ManageYearGroupsModal';
+import { ManageCohortsModal } from '../referenceData/ManageCohortsModal';
+import { ManageYearGroupsModal } from '../referenceData/ManageYearGroupsModal';
 import { bulkSetCohort, getActiveCohortOptions } from './bulk/bulkSetCohortFlow';
 import { bulkSetCourseLength } from './bulk/bulkSetCourseLengthFlow';
 import { bulkSetYearGroup, getYearGroupOptions } from './bulk/bulkSetYearGroupFlow';
