+++ src/frontend/src/features/referenceData/ManageCohortsModal.tsx
@@ -15,12 +15,12 @@
 
 import { Button, Space, Switch, type TableColumnType } from 'antd';
 import type { UseQueryOptions } from '@tanstack/react-query';
-import type { Cohort } from '../../../services/referenceData/referenceData.zod';
-import { createCohort, deleteCohort, updateCohort } from '../../../services/referenceData/referenceDataService';
-import { getCohortsQueryOptions } from '../../../query/sharedQueries';
-import { ReferenceDataInitialLoadingState } from '../components/ReferenceDataInitialLoadingState';
+import type { Cohort } from '../../services/referenceData/referenceData.zod';
+import { createCohort, deleteCohort, updateCohort } from '../../services/referenceData/referenceDataService';
+import { getCohortsQueryOptions } from '../../query/sharedQueries';
+import { ReferenceDataInitialLoadingState } from '../classes/components/ReferenceDataInitialLoadingState';
 import { ReferenceDataManagementModalScaffold } from './ReferenceDataManagementModalScaffold';
-import { useReferenceDataManagement } from '../hooks/useReferenceDataManagement';
+import { useReferenceDataManagement } from '../classes/hooks/useReferenceDataManagement';
 
 export type ManageCohortsModalProperties = Readonly<{
   open: boolean;
