+++ src/frontend/src/features/referenceData/ManageYearGroupsModal.tsx
@@ -15,16 +15,16 @@
 
 import { Button, Space, type TableColumnType } from 'antd';
 import type { UseQueryOptions } from '@tanstack/react-query';
-import type { YearGroup } from '../../../services/referenceData/referenceData.zod';
+import type { YearGroup } from '../../services/referenceData/referenceData.zod';
 import {
   createYearGroup,
   deleteYearGroup,
   updateYearGroup,
-} from '../../../services/referenceData/referenceDataService';
-import { getYearGroupsQueryOptions } from '../../../query/sharedQueries';
-import { ReferenceDataInitialLoadingState } from '../components/ReferenceDataInitialLoadingState';
+} from '../../services/referenceData/referenceDataService';
+import { getYearGroupsQueryOptions } from '../../query/sharedQueries';
+import { ReferenceDataInitialLoadingState } from '../classes/components/ReferenceDataInitialLoadingState';
 import { ReferenceDataManagementModalScaffold } from './ReferenceDataManagementModalScaffold';
-import { useReferenceDataManagement } from '../hooks/useReferenceDataManagement';
+import { useReferenceDataManagement } from '../classes/hooks/useReferenceDataManagement';
 
 export type ManageYearGroupsModalProperties = Readonly<{
   open: boolean;
