+++ src/frontend/src/features/referenceData/ReferenceDataManagementModalScaffold.tsx
