+++ src/frontend/src/features/referenceData/manageCohorts.spec.tsx
@@ -9,18 +9,18 @@
 import { act, fireEvent, screen, waitFor, within } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { QueryClientProvider } from '@tanstack/react-query';
-import type { Cohort } from '../../../services/referenceData/referenceData.zod';
-import { queryKeys } from '../../../query/queryKeys';
-import { createAppQueryClient } from '../../../query/queryClient';
-import { renderWithFrontendProviders } from '../../../test/renderWithFrontendProviders';
-import { StartupWarmupStateProvider } from '../../../features/auth/startupWarmupState';
+import type { Cohort } from '../../services/referenceData/referenceData.zod';
+import { queryKeys } from '../../query/queryKeys';
+import { createAppQueryClient } from '../../query/queryClient';
+import { renderWithFrontendProviders } from '../../test/renderWithFrontendProviders';
+import { StartupWarmupStateProvider } from '../../features/auth/startupWarmupState';
 import { ManageCohortsModal } from './ManageCohortsModal';
 
 const createCohortMock = vi.hoisted(() => vi.fn());
 const updateCohortMock = vi.hoisted(() => vi.fn());
 const getCohortsMock = vi.hoisted(() => vi.fn());
 
-vi.mock('../../../services/referenceData/referenceDataService', () => ({
+vi.mock('../../services/referenceData/referenceDataService', () => ({
   getCohorts: getCohortsMock,
   createCohort: createCohortMock,
   updateCohort: updateCohortMock,
