+++ src/frontend/src/features/referenceData/manageYearGroups.spec.tsx
@@ -11,18 +11,18 @@
 import { act, fireEvent, screen, waitFor, within } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { QueryClientProvider } from '@tanstack/react-query';
-import type { YearGroup } from '../../../services/referenceData/referenceData.zod';
-import { queryKeys } from '../../../query/queryKeys';
-import { createAppQueryClient } from '../../../query/queryClient';
-import { renderWithFrontendProviders } from '../../../test/renderWithFrontendProviders';
-import { StartupWarmupStateProvider } from '../../../features/auth/startupWarmupState';
+import type { YearGroup } from '../../services/referenceData/referenceData.zod';
+import { queryKeys } from '../../query/queryKeys';
+import { createAppQueryClient } from '../../query/queryClient';
+import { renderWithFrontendProviders } from '../../test/renderWithFrontendProviders';
+import { StartupWarmupStateProvider } from '../../features/auth/startupWarmupState';
 import { ManageYearGroupsModal } from './ManageYearGroupsModal';
 
 const createYearGroupMock = vi.hoisted(() => vi.fn());
 const updateYearGroupMock = vi.hoisted(() => vi.fn());
 const getYearGroupsMock = vi.hoisted(() => vi.fn());
 
-vi.mock('../../../services/referenceData/referenceDataService', () => ({
+vi.mock('../../services/referenceData/referenceDataService', () => ({
   getCohorts: vi.fn(),
   createCohort: vi.fn(),
   updateCohort: vi.fn(),
