+++ src/frontend/src/pages/SettingsPage.tsx
@@ -3,7 +3,7 @@ import type { TabsProps } from 'antd';
 import { useMemo, useState } from 'react';
 import { ClassesManagementPanel } from '../features/classes/ClassesManagementPanel';
 import { BackendSettingsPanel } from '../features/settings/backend/BackendSettingsPanel';
-import { ReferenceDataSettingsPanel } from '../features/settings/ReferenceDataSettingsPanel';
+import { ReferenceDataSettingsPanel } from '../features/referenceData/ReferenceDataSettingsPanel';
 import { PageSection } from './PageSection';
 import { SettingsPageGoogleClassroomsPrefetch } from '../features/settings/SettingsPageGoogleClassroomsPrefetch';
 import { pageContent } from './pageContent';
